Type Alias pallet_asset_switch::pallet::Module

pub type Module<T, I = ()> = Pallet<T, I>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T, I = ()>(/* private fields */);

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>where LocalCurrencyBalanceOf<T, I>: Into<u128>,

pub fn set_switch_pair( origin: OriginFor<T>, remote_asset_total_supply: u128, remote_asset_id: Box<VersionedAssetId>, remote_asset_circulating_supply: u128, remote_reserve_location: Box<VersionedLocation>, remote_asset_ed: u128, remote_xcm_fee: Box<VersionedAsset> ) -> DispatchResult

Set a new switch pair.

See the crate’s README for more.

pub fn force_set_switch_pair( origin: OriginFor<T>, remote_asset_total_supply: u128, remote_asset_id: Box<VersionedAssetId>, remote_asset_circulating_supply: u128, remote_reserve_location: Box<VersionedLocation>, remote_asset_ed: u128, remote_xcm_fee: Box<VersionedAsset> ) -> DispatchResult

Force-set a new switch pair.

See the crate’s README for more.

pub fn force_unset_switch_pair(origin: OriginFor<T>) -> DispatchResult

Unset a switch pair.

See the crate’s README for more.

pub fn pause_switch_pair(origin: OriginFor<T>) -> DispatchResult

Pause switches for a switch pair.

See the crate’s README for more.

pub fn resume_switch_pair(origin: OriginFor<T>) -> DispatchResult

Resume switches for a switch pair.

See the crate’s README for more.

pub fn update_remote_xcm_fee( origin: OriginFor<T>, new: Box<VersionedAsset> ) -> DispatchResult

Update the remote XCM fee for a switch pair.

See the crate’s README for more.

pub fn switch( origin: OriginFor<T>, local_asset_amount: LocalCurrencyBalanceOf<T, I>, beneficiary: Box<VersionedLocation> ) -> DispatchResult

Perform a local -> remote asset switch.

See the crate’s README for more.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn switch_pair() -> Option<SwitchPairInfoOf<T>>

An auto-generated getter for SwitchPair.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn pending_switch_confirmations<KArg>( k: KArg ) -> Option<UnconfirmedSwitchInfoOf<T>>where KArg: EncodeLike<QueryId>,

An auto-generated getter for PendingSwitchConfirmations.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn next_query_id() -> QueryId

An auto-generated getter for NextQueryId.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn pool_account_id_for_remote_asset( remote_asset_id: &VersionedAssetId ) -> Result<T::AccountId, Error<T, I>>

Derive an AccountId for the provided remote_asset_id and the pallet’s name as configured in the runtime.

pub fn compute_xcm_for_switch( inverted_universal_location: &Location, from: &Location, to: &Location, amount: u128, asset_id: &AssetId, remote_asset_fee: &Asset ) -> Xcm<()>

Compose the XCM message for a local -> remote switch to be sent to the switch pair configured remote location.

The message includes an appendix which reports back to this chain the result of the operation by the means of sending the content of the holding registry after the transfer is supposed to have happened.

The XCM program is so composed:

Withdraw XCM fees from our SA

Buy execution

Set the appendix, executed regardless of the outcome of the transfer:

3.1 Report back to our chain the assets in the holding registry. This will contain either only the XCM fee token in case of successful transfer, or the XCM fee token + the amount of funds supposed to be transferred.

3.2 Deposit the un-transferred asset (only if the transfer failed) back into our account.

3.3 Refund any surplus weight.

3.4 Deposit the remaining XCM fee assets in the user’s account.

Withdraw the requested asset (this operation should be infallible since we have full control of this balance)

Try to deposit the withdrawn asset into the user’s account. This operation could fail and the error is handled in the appendix.