Type Alias pallet_asset_switch::pallet::NewSwitchPairInfoOf

pub type NewSwitchPairInfoOf<T> = NewSwitchPairInfo<<T as Config>::AccountId>;

Aliased Type

struct NewSwitchPairInfoOf<T> {
    pub pool_account: <T as Config>::AccountId,
    pub remote_asset_circulating_supply: u128,
    pub remote_asset_ed: u128,
    pub remote_asset_id: VersionedAssetId,
    pub remote_asset_total_supply: u128,
    pub remote_reserve_location: VersionedLocation,
    pub remote_xcm_fee: VersionedAsset,
    pub status: SwitchPairStatus,
}

Fields
pool_account: <T as Config>::AccountId

The address that will hold the local tokens locked in return for the remote asset.

remote_asset_circulating_supply: u128

The circulating supply, i.e., the total supply - required EDs for both local and remote assets - supply controlled by the chain on the remote reserve location.

remote_asset_ed: u128

The existential deposit (i.e., minimum balance to hold) of the remote asset.

remote_asset_id: VersionedAssetId

The ID of the remote asset to switch 1:1 with the local token.

remote_asset_total_supply: u128

The total supply of the remote asset. This is assumed to never change. If it does, the current pool must be manually updated to reflect the changes.

remote_reserve_location: VersionedLocation

The remote location on which the remote asset lives.

remote_xcm_fee: VersionedAsset

The assets to take from the user’s balance on this chain to pay for XCM execution fees on the reserve location.

status: SwitchPairStatus

The status of the switch pair.