Trait pallet_asset_switch::traits::SwitchHooks

pub trait SwitchHooks<T, I>where
    T: Config<I>,
    I: 'static,{
    type Error: Into<u8>;

    // Required methods
    fn pre_local_to_remote_switch(
        from: &T::AccountId,
        to: &VersionedLocation,
        amount: LocalCurrencyBalanceOf<T, I>
    ) -> Result<(), Self::Error>;
    fn post_local_to_remote_switch_dispatch(
        from: &T::AccountId,
        to: &VersionedLocation,
        amount: LocalCurrencyBalanceOf<T, I>
    ) -> Result<(), Self::Error>;
    fn post_local_to_remote_finalized(
        from: &T::AccountId,
        to: &VersionedLocation,
        amount: LocalCurrencyBalanceOf<T, I>
    );
    fn post_local_to_remote_transfer_revert(
        from: &T::AccountId,
        to: &VersionedLocation,
        amount: LocalCurrencyBalanceOf<T, I>
    );
    fn pre_remote_to_local_switch(
        to: &T::AccountId,
        amount: u128
    ) -> Result<(), Self::Error>;
    fn post_remote_to_local_switch(
        to: &T::AccountId,
        amount: u128
    ) -> Result<(), Self::Error>;
}

Runtime-injected logic into the switch pallet around the time a switch takes place.

The order in which the hooks are called, for any given local -> remote switch is the following:

pre_local_to_remote_switch (fallible): Called from within the switch pallet before any other logic is executed. E.g., it can be used to verify some requirements about the transfer.

post_local_to_remote_switch_dispatch (fallible): Called from within the switch pallet after all other logic is executed and after the XCM message for the transfer is prepared to be sent to the remote location. E.g., it can be used to update storage elements of other pallets once the transfer operation is guaranteed to be sent to destination. This function can still fail, in which case the message won’t be sent. But otherwise, at this stage the XCM message is guaranteed to at least be deliverable to the destination. It can still fail at destination, which will in turn call the post_local_to_remote_transfer_revert hook.

post_local_to_remote_confirmed (infallible): Called when the remote destination confirms the transfer was successful.

post_local_to_remote_transfer_revert (infallible): Called when the remote destination signals that a previously local -> remote switch has instead failed.

The order in which the hooks are called, for any given remote -> local switch is the following:

pre_remote_to_local_switch (fallible): Called from the XCM components handling incoming transfers representing remote -> local switches. If this hook fails, the assets will most likely trap and need to be unlocked manually.

post_remote_to_local_switch (fallible): Same as pre_remote_to_local_switch, but called after all the XCM logic has been executed and guaranteed to be correct.

Required Associated Types

type Error: Into<u8>

Required Methods

fn pre_local_to_remote_switch( from: &T::AccountId, to: &VersionedLocation, amount: LocalCurrencyBalanceOf<T, I> ) -> Result<(), Self::Error>

fn post_local_to_remote_switch_dispatch( from: &T::AccountId, to: &VersionedLocation, amount: LocalCurrencyBalanceOf<T, I> ) -> Result<(), Self::Error>

fn post_local_to_remote_finalized( from: &T::AccountId, to: &VersionedLocation, amount: LocalCurrencyBalanceOf<T, I> )

fn post_local_to_remote_transfer_revert( from: &T::AccountId, to: &VersionedLocation, amount: LocalCurrencyBalanceOf<T, I> )

fn pre_remote_to_local_switch( to: &T::AccountId, amount: u128 ) -> Result<(), Self::Error>

fn post_remote_to_local_switch( to: &T::AccountId, amount: u128 ) -> Result<(), Self::Error>