Trait pallet_asset_switch::traits::SwitchHooks

pub trait SwitchHooks<T, I>where
    T: Config<I>,
    I: 'static,{
    type Error: Into<u8>;

    // Required methods
    fn pre_local_to_remote_switch(
        from: &T::AccountId,
        to: &VersionedLocation,
        amount: LocalCurrencyBalanceOf<T, I>
    ) -> Result<(), Self::Error>;
    fn post_local_to_remote_switch(
        from: &T::AccountId,
        to: &VersionedLocation,
        amount: LocalCurrencyBalanceOf<T, I>
    ) -> Result<(), Self::Error>;
    fn pre_remote_to_local_switch(
        to: &T::AccountId,
        amount: u128
    ) -> Result<(), Self::Error>;
    fn post_remote_to_local_switch(
        to: &T::AccountId,
        amount: u128
    ) -> Result<(), Self::Error>;
}

Runtime-injected logic into the switch pallet around the time a switch takes place.

Required Associated Types

type Error: Into<u8>

Required Methods

fn pre_local_to_remote_switch( from: &T::AccountId, to: &VersionedLocation, amount: LocalCurrencyBalanceOf<T, I> ) -> Result<(), Self::Error>

Called before anything related to a switch happens.

fn post_local_to_remote_switch( from: &T::AccountId, to: &VersionedLocation, amount: LocalCurrencyBalanceOf<T, I> ) -> Result<(), Self::Error>

Called after the switch takes place and after the XCM message has been sent to the reserve location.

fn pre_remote_to_local_switch( to: &T::AccountId, amount: u128 ) -> Result<(), Self::Error>

Called upon receiving an XCM message from the reserve location to deposit some of the remote assets into a specified account, but before the asset is actually deposited by the asset transactor.

fn post_remote_to_local_switch( to: &T::AccountId, amount: u128 ) -> Result<(), Self::Error>

Same as Self::pre_remote_to_local_switch, but called after the transactor has deposited the incoming remote asset.